add missing order checks, user-actions file
